requirements.txt
pip>=24.0
setuptools>=70.0
wheel>=0.43
alembic==1.14.1
aniso8601==9.0.1
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
distlib==0.3.9
dominate==2.9.1
filelock==3.17.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
numpy==1.26.0
packaging==24.1
pandas==2.0.3
phonenumbers==8.13.37
platformdirs==4.3.6
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2024.1
sendgrid==6.11.0
six==1.16.0
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.29.1
visitor==0.1.3
Werkzeug==3.1.3
